+++ src/pages/home/home.ts
@@ -5,6 +5,7 @@ import { BehaviorSubject, Observable, Subscription } from 'rxjs';
 import { LocationsPage } from '../locations/locations';
 
 import { LocationsProvider } from '../../providers/locations/locations';
+import { GeolocationProvider } from '../../providers/geolocation/geolocation';
 import { HttpProvider } from '../../providers/http/http';
 
 import { Geolocation } from '@ionic-native/geolocation';
@@ -22,69 +23,13 @@ export class HomePage {
   public locationObj;
   public latestLocations;
 
-  constructor(public navCtrl: NavController, public geolocation: Geolocation, public locationsProvider: LocationsProvider, public httpProvider: HttpProvider, 
+  constructor(public navCtrl: NavController, public geolocation: Geolocation, public locationsProvider: LocationsProvider, public httpProvider: HttpProvider, public geolocationProvider: GeolocationProvider, 
               private camera: Camera, private alertCtrl: AlertController) {
     
   }
 
   ionViewDidLoad() {
-    this.getCurrentLocation();
-  }
-
-  // Opções para ativar alta precisão na hora de retornar coordenadas
-  locationOptions = {
-    enableHighAccuracy: true
-  }
-  
-  // Pega as coordenadas atuais quando a view é carregada e começa a chamar a função que irá retornar
-  // as coordenadas a cada três minutos
-  getCurrentLocation() {
-    this.getCurrentTime();
-    this.geolocation.getCurrentPosition(this.locationOptions).then((resp) => {
-      this.watchPosition(); // função pra continuar loop de verificação de coords
-      this.lat = resp.coords.latitude;
-      this.lng = resp.coords.longitude;
-      this.locationObj = {
-        time: this.timedate,
-        lat: this.lat,
-        lng: this.lng
-      }
-      // Insere localização no banco
-      this.locationsProvider.insertLocations(this.locationObj)
-        .then(res => { 
-          // Pega localizações após inserir
-          this.locationsProvider.getAllLocations()
-            .then(res => {
-              // Trata array após pegar localizações
-              this.getLatestLocations() 
-            }
-          );
-        }
-      );
-    }).catch((error) => {
-      console.log('Error getting location', error);
-    });
-  }
-
-  //Pega as três ultimas localizações da base para exibir na home
-  getLatestLocations() {
-    this.latestLocations = this.locationsProvider.locations.slice(0, 3);
-  }
-
-  // Timeout de 3 minutos pra chamar novamente a função que pega as coordenadas
-  watchPosition() {
-    setTimeout(() => {
-      this.getCurrentLocation();
-    }, 
-    180000);
-  }
-
-  // Pega hora e minutos
-  getCurrentTime() {
-    let todayDate = new Date();
-    let hour = todayDate.getHours();
-    let minutes = todayDate.getMinutes();
-    minutes < 10 ? this.timedate = hour.toString() + ":0" + minutes.toString() : this.timedate = hour.toString() + ":" + minutes.toString()
+    this.geolocationProvider.getCurrentLocation();
   }
 
   //Envia localizações e imagem tirada com camera para o servidor
